# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/events/register/{event_id}": {
            "post": {
                "tags": [
                    "Event Registrations"
                ],
                "summary": "Register for an event",
                "operationId": "88d63af8bc191e0139aa02ea7b8eda4c",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID of the event to register for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully registered for the event",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventRegistration"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You already registered for this event, No places available, event is full"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/event-registrations/{id}": {
            "get": {
                "tags": [
                    "Event Registrations"
                ],
                "summary": "Get a specific event registration",
                "operationId": "76aac0839e6986905beefef0638fa159",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event registration to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventRegistration"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event registration not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event Registration not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Event Registrations"
                ],
                "summary": "Update an event registration",
                "operationId": "681035147f3caac8f27e56270b732d7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event registration to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "event_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "l6k7j8h9-0123-45gf-dsaq-wertyuioplkj"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "some-user-uuid"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "approved",
                                            "rejected"
                                        ],
                                        "example": "approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event registration updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventRegistration"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event registration not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event Registration not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Event Registrations"
                ],
                "summary": "Delete an event registration",
                "operationId": "f0884cc615843408a062b663b80e9f69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event registration to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event registration deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event Registration deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event registration not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event Registration not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EventRegistration": {
                "title": "EventRegistration",
                "description": "Event Registration model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the event registration",
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "event_id": {
                        "description": "ID of the event being registered for",
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "description": "ID of the user registering for the event",
                        "type": "string",
                        "format": "uuid"
                    },
                    "registered_at": {
                        "description": "Timestamp when the user registered for the event",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "status": {
                        "description": "Status of the event registration",
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ],
                        "example": "pending"
                    },
                    "created_at": {
                        "description": "Timestamp when the event registration was created",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Timestamp when the event registration was last updated",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EventRegistration"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}